RoboCommandPortable is now passing all unit tests locally.
- Consolidated various test cases (for example, ones that were passing and tested for already in CommandTests<T>
- Tests should be using random directories now instead of a shared destination, which could potentially create conflicts.
- Removed TestPrep.cs since it is no longer relevant.
- Adjusted extensions as needed to get all tests passing and aligned with actual robocopy functionality.

+++ RoboSharp.Extensions.UnitTests/Windows/SymbolicLinkTests.cs
@@ -1,13 +1,10 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿#if WINDOWS
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RoboSharp.Extensions.SymbolicLinkSupport;
-using RoboSharp.UnitTests;
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Windows.Win32.Storage.FileSystem;
+using System.Security.Principal;
 
 
 namespace RoboSharp.Extensions.Windows.UnitTests
@@ -25,6 +22,11 @@ namespace RoboSharp.Extensions.Windows.UnitTests
         [TestInitialize]
         public void Initialize_SymbolicLinkTest()
         {
+            if (!IsAdministrator())
+            {
+                Console.WriteLine("Test Requires Admin Privileges");
+                return;
+            }
             Console.WriteLine("----------------------\n THIS TEST REQUIRES ADMIN PRIVILEGES (OR DEVELOPER-MODE ENABLED) AT RUN-TIME ");
             RoboSharp.UnitTests.Test_Setup.PrintEnvironment();
             Directory.CreateDirectory(Root);
@@ -32,15 +34,30 @@ namespace RoboSharp.Extensions.Windows.UnitTests
             SymbolicLink.ALLOW_UNPRIVILEGED_CREATE = true;
         }
 
+
+        private static bool IsAdministrator()
+        {
+            using (WindowsIdentity identity = WindowsIdentity.GetCurrent())
+            {
+                WindowsPrincipal principal = new WindowsPrincipal(identity);
+                return principal.IsInRole(WindowsBuiltInRole.Administrator);
+            }
+        }
+
         [TestCleanup]
         public void Cleanup_SymbolicLinkTest()
         {
+            if (!IsAdministrator())
+                return;
             Directory.Delete(Root, true);
         }
 
         [TestMethod]
         public void Test_SymbolicDirectory()
         {
+            if (!IsAdministrator()) 
+                return;
+
             DirectoryInfo target = new DirectoryInfo(Path.Combine(TargetRoot, RandomName(), RandomName()));
             DirectoryInfo link = new DirectoryInfo(Path.Combine(Root, RandomName()));
 
@@ -99,6 +116,9 @@ namespace RoboSharp.Extensions.Windows.UnitTests
         [TestMethod]
         public void Test_SymbolicFile()
         {
+            if (!IsAdministrator())
+                return;
+
             FileInfo target = new FileInfo(Path.Combine(TargetRoot, Path.GetFileNameWithoutExtension(Path.GetRandomFileName()) + ".txt"));
             FileInfo link = new FileInfo(Path.Combine(Root, target.Name));
             try
@@ -174,3 +194,5 @@ namespace RoboSharp.Extensions.Windows.UnitTests
 
     }
 }
+
+#endif
